Add `EdgeBuilder`
