+++ crates/fj-kernel/src/builder/mod.rs
@@ -1,9 +1,11 @@
 //! API for building objects
 
 mod cycle;
+mod edge;
 mod face;
 
 pub use self::{
     cycle::CycleBuilder,
+    edge::EdgeBuilder,
     face::{FaceBuilder, FacePolygon},
 };
